requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.9
html5lib==1.1
idna==3.3
install==1.3.5
lxml==4.7.1
python-dotenv==0.19.2
requests==2.27.0
six==1.16.0
soupsieve==2.3.1
spotipy==2.19.0
urllib3==1.26.7
webencodings==0.5.1
